cpp reverse bits

+++ cpp/example/src/ReverseBits/reverseBits.h
@@ -4,5 +4,5 @@ using namespace std;
 
 class reverseBits{
     public:
-        static int naive(int n);
+        static long naive(long n);
 };
